timApp/modules/chattim/requirements.txt
Flask
Flask-WTF
marshmallow==3.22.0
marshmallow-enum==1.5.1
marshmallow-union==0.1.15.post1
webargs==5.5.0
typing-inspect==0.9.0
requests
python-dateutil
isodate
gunicorn==21.2.0
gevent>=24.10.1
openai
google-genai
